Bug fixes from error logs

+++ app/Http/Controllers/ConsultationController.php
@@ -6,7 +6,6 @@ use App\Models\Consultation;
 use App\Http\Requests\StoreConsultationRequest;
 use App\Http\Requests\StoreConsultationReviewRequest;
 use App\Http\Requests\UpdateAdmissionStatusRequest;
-use App\Http\Requests\UpdateConsultationRequest;
 use App\Http\Resources\ConsultationReviewCollection;
 use App\Http\Resources\LatestLmpResource;
 use App\Http\Resources\PatientBioResource;
@@ -72,6 +71,9 @@ class ConsultationController extends Controller
     public function destroy(Consultation $consultation)
     {
        return DB::transaction(function () use ($consultation) {
+                    if ($consultation->prescriptions->count() > 0){
+                        return response()->json(['message' => 'Delete prescriptions first'], 222);
+                    }
                     $consultation->visit->consultations->count() < 2 ? $consultation->visit->update(['consulted' => null]) : '' ;
                     $consultation->destroy($consultation->id);
                 }, 2);
